Split lines from file...
For each domain we want to be able to set it to a different address,
this would allow for example:

apple.local:apple.com

And a CNAME record for apple.local would be set to apple.com

We also allow $self to be used. This retrieves the current hostname
using a gethostname() call. This way you can create files that can be
used from system to system.