Add Registration Domain lookup
This fetches all of the registration domains that are available, the
goal is to not start registrating all of our CNAME's until we know that
it is a registration domain that is available.

This will call a start registration callback that passes the domain
that was just retrieved, this can then be used to start registering
CNAMES within that domain.